requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
certifi==2024.12.14
click==8.1.8
colorama==0.4.6
distro==1.9.0
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.60.0
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
python-telegram-bot==21.10
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uvicorn==0.34.0
